% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cat.R
\name{calc_cat}
\alias{calc_cat}
\title{Calculate Summary Statistics for Categorical Variables}
\usage{
calc_cat(d, var)
}
\arguments{
\item{d, }{var}
}
\description{
Calculate Summary Statistics for Categorical Variables
}
\keyword{table}

