% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_contin.R
\name{calc_contin}
\alias{calc_contin}
\title{Calculate Summary Statistics for Continuous Variables}
\usage{
calc_contin(d, var)
}
\arguments{
\item{d, }{var}
}
\description{
Calculate Summary Statistics for Continuous Variables
}
\keyword{table}

