% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_sd.R
\name{fun_miss}
\alias{fun_miss}
\title{Fun_miss
A function that takes a vector as an argument and returns the number of Missing variables it has}
\usage{
fun_miss(z)
}
\arguments{
\item{z}{is a vector with a number of elements}
}
\value{
Returns a count of the Missing Variables
}
\description{
Fun_miss
A function that takes a vector as an argument and returns the number of Missing variables it has
}
\examples{
y <- c(rep(c(45:60, 70:80, NA), 10), 400, 200, 9, rep(c(NA), 20))
fun_miss(z)
}
