% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_mean.R
\name{my_mean}
\alias{my_mean}
\title{my_mean function}
\usage{
my_mean(y, method_na)
}
\arguments{
\item{y}{is a vector with a number of elements}

\item{method_na}{uses a method to impute Missing Values with the mean of non NA vector values}
}
\value{
Returns a mean of a vector with Missing Values imputed with the mean of non NA vector values
}
\description{
my_mean function
}
\examples{
y <- c(rep(c(45:60, 70:80, NA), 10), 400, 200, 9, rep(c(NA), 20))
my_mean(y, method_na = "impute")

}
