% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_miss.R
\name{my_sd}
\alias{my_sd}
\title{my_sd Function}
\usage{
my_sd(n)
}
\arguments{
\item{n}{is a vector of elements}
}
\value{
Standard deviation after imputing for Outliers and NAs with mean of elements without NAs
}
\description{
my_sd Function
}
\examples{
x<- c(rep(c(45:60, 70:80, NA), 10), 400, 200, 9, rep(c(NA), 20))
my_sd(x)
}
