% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{sum_stats}
\alias{sum_stats}
\title{Calculate Summary Statistics by Group}
\usage{
sum_stats(data, sort_var, stat_var)
}
\arguments{
\item{data}{Input data.}

\item{sort_var}{Variable used for grouping; named as sort_var to indicate that it is used to sort
the data into groups for subsequent analysis}

\item{stat_var}{Numeric variable which is used to calculate summary statistics. Named stat_var as
this is the variable with which the summary statistics are calculated}
}
\value{
A tibble is returned which contains the grouping variable and summary statistics as columns (min, max, mean, and standard deviation)
}
\description{
The function calculates four summary statistics (min, max, mean, standard deviation) of a variable
by group within a data table. The variables used for grouping and summarizing are input by the user.
}
\examples{
if (requireNamespace("palmerpenguins", quietly = TRUE)) {
  penguins <- palmerpenguins::penguins
  sum_stats(penguins, island, body_mass_g)
}

if (requireNamespace("datateachr", quietly = TRUE)) {
  cancer_sample <- datateachr::cancer_sample
  sum_stats(cancer_sample, diagnosis, area_mean)
}
}
